* 'developer' of https://github.com/zhiyicx/thinksns-plus-android: (31 commits)
  (#103) test 添加6.0敏感权限文档
  (#103) code 添加6.0敏感权限适配
  (# 76) code 添加发送动态到动态列表的状态判定，更改发送动态后台任务逻辑
  （# 76） code DynamicBean添加三种发送状态
  (# 76 ) code  将动态发送状态存在DynamicBean表中
  (# 76) code 添加动态数据库的Dagger管理
  （# 76） code 添加动态数据库操作方法
  (# 76) code 添加动态Bean的greenDao
  (#86) code 修改基础类别 max—_id 类型为Longg
  （#7）code 添加recyclerVIew的exspresso框架
  （# 76） code 添加发送动态测试
  (#86) code 获取网络动态列表数据
  (#86) code 添加获取动态列表接口
  （# 76）code 修改图片判断为：没有图片或者只有占位图时，不可点击
  (#86) code 动态数据处理基类
  (#87) test 修改音乐fm测试
  doc 修改baseadapter 使用说明　
  (# 64) code 调整相册列表布局
  (# 64) code 调整相册的布局
  （# 76）code 添加发送动态的测试类
  ...

+++ app/src/main/java/com/zhiyicx/thinksnsplus/data/source/repository/FollowFansListRepository.java
@@ -3,7 +3,6 @@ package com.zhiyicx.thinksnsplus.data.source.repository;
 import com.zhiyicx.common.base.BaseJson;
 import com.zhiyicx.thinksnsplus.data.beans.FollowFansBean;
 import com.zhiyicx.thinksnsplus.data.beans.GsonFollowFansBean;
-import com.zhiyicx.thinksnsplus.data.beans.UserInfoBean;
 import com.zhiyicx.thinksnsplus.data.source.remote.FollowFansClient;
 import com.zhiyicx.thinksnsplus.data.source.remote.ServiceManager;
 import com.zhiyicx.thinksnsplus.modules.follow_fans.FollowFansListContract;
@@ -12,10 +11,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import rx.Observable;
-import rx.Subscriber;
-import rx.android.schedulers.AndroidSchedulers;
 import rx.functions.Func1;
-import rx.schedulers.Schedulers;
 
 /**
  * @author LiuChao
@@ -42,7 +38,7 @@ public class FollowFansListRepository implements FollowFansListContract.Reposito
                         List<FollowFansBean> followFansBeanList = new ArrayList<FollowFansBean>();
                         for (GsonFollowFansBean.GsonFollowsBean gsonFollowsBean : gsonFollowFansBean.getFollows()) {
                             FollowFansBean followFansBean = new FollowFansBean();
-                            followFansBean.setMaxId(gsonFollowsBean.getId());// 存入maxId
+                            followFansBean.setMaxId((long) gsonFollowsBean.getId());// 存入maxId
                             // 关注主体是当前传入的userId
                             followFansBean.setUserId(userId);
                             followFansBean.setFollowedUserId(gsonFollowsBean.getUser_id());
@@ -71,7 +67,7 @@ public class FollowFansListRepository implements FollowFansListContract.Reposito
                         List<FollowFansBean> followFansBeanList = new ArrayList<FollowFansBean>();
                         for (GsonFollowFansBean.GsonFollowsBean gsonFollowsBean : gsonFollowFansBean.getFolloweds()) {
                             FollowFansBean followFansBean = new FollowFansBean();
-                            followFansBean.setMaxId(gsonFollowsBean.getId());// 存入maxId
+                            followFansBean.setMaxId((long) gsonFollowsBean.getId());// 存入maxId
                             // 关注主体是当前从服务器获取到的user_id
                             followFansBean.setUserId(gsonFollowsBean.getUser_id());
                             followFansBean.setFollowedUserId(userId);
